requirements.txt
pillow~=5.1.0
xmltodict~=0.12.0
requests~=2.25.1
beautifulsoup4~=4.9.3
PyPDF2~=1.28.2